Refactor HideDamageSplash to be in combat
